% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_exists_conda}
\alias{is_exists_conda}
\title{Is conda installed?}
\usage{
is_exists_conda()
}
\value{
Logical, true if and only if conda is installed.
}
\description{
Is conda installed?
}
\author{
Resul Akay
}
