% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ngboost}
\alias{ngboost}
\title{Probabilistic forecasting using Python's NGBoost library.}
\description{
The goal of ngboostForecast is to provide a probabilistic forecasting
interface for Python's \href{https://stanfordmlgroup.github.io}{NGBoost} library
}
