% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{seatbelts}
\alias{seatbelts}
\title{Road Casualties in Great Britain 1969-84}
\format{
An object of class \code{mts} (inherits from \code{ts}) with 192 rows and 8 columns.
}
\source{
Harvey, A.C. (1989). Forecasting, Structural Time Series Models and the
Kalman Filter. Cambridge University Press, pp. 519–523.

Durbin, J. and Koopman, S. J. (2001). Time Series Analysis by State Space
Methods. Oxford University Press.

\url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/UKDriverDeaths.html}
}
\usage{
seatbelts
}
\description{
The Seatbelts dataset from the datasets package.
}
\references{
Harvey, A. C. and Durbin, J. (1986). The effects of seat belt legislation on
British road casualties: A case study in structural time series modelling.
Journal of the Royal Statistical Society series A, 149, 187–227.
}
\keyword{datasets}
