% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sklearner}
\alias{sklearner}
\title{Scikit-Learn interface}
\usage{
sklearner(module = "tree", class = "DecisionTreeRegressor", ...)
}
\arguments{
\item{module}{scikit-learn module name, default is 'tree'.}

\item{class}{scikit-learn's module class, default is 'DecisionTreeRegressor'}

\item{...}{Other arguments passed to model class}
}
\description{
Scikit-Learn interface
}
\examples{
\dontrun{

sklearner(module = "tree", class = "DecisionTreeRegressor",
criterion="friedman_mse", min_samples_split=2)

}

}
\author{
Resul Akay
}
